refactiring item entries
