+++ src/VisualStudio/Core/Def/Implementation/FindReferences/Entries/AbstractItemEntry.cs
@@ -1,0 +1,41 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using System.Windows;
+using System.Windows.Documents;
+using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
+using Microsoft.VisualStudio.Shell.TableControl;
+
+namespace Microsoft.VisualStudio.LanguageServices.FindUsages
+{
+    internal partial class StreamingFindUsagesPresenter
+    {
+        private abstract class AbstractItemEntry : Entry
+        {
+            protected readonly StreamingFindUsagesPresenter _presenter;
+
+            public AbstractItemEntry(RoslynDefinitionBucket definitionBucket, StreamingFindUsagesPresenter presenter)
+                : base(definitionBucket)
+            {
+                _presenter = presenter;
+            }
+
+            public override bool TryCreateColumnContent(string columnName, out FrameworkElement content)
+            {
+                if (columnName == StandardTableColumnDefinitions2.LineText)
+                {
+                    var inlines = CreateLineTextInlines();
+                    var textBlock = inlines.ToTextBlock(_presenter.ClassificationFormatMap, wrap: false);
+
+                    content = textBlock;
+                    return true;
+                }
+
+                content = null;
+                return false;
+            }
+
+            protected abstract IList<Inline> CreateLineTextInlines();
+        }
+    }
+}
