+++ src/VisualStudio/Core/Def/Implementation/FindReferences/Entries/DocumentSpanEntry.cs
@@ -69,7 +69,7 @@ namespace Microsoft.VisualStudio.LanguageServices.FindUsages
                         ? WrittenReferenceHighlightTag.TagId
                         : ReferenceHighlightTag.TagId;
 
-                var properties = Presenter.FormatMapService
+                var properties = _presenter.FormatMapService
                                           .GetEditorFormatMap("text")
                                           .GetProperties(propertyId);
 
@@ -83,8 +83,8 @@ namespace Microsoft.VisualStudio.LanguageServices.FindUsages
                     cs => new ClassifiedText(cs.ClassificationType, _excerptResult.Content.ToString(cs.TextSpan)));
 
                 var inlines = classifiedTexts.ToInlines(
-                    Presenter.ClassificationFormatMap,
-                    Presenter.TypeMap,
+                    _presenter.ClassificationFormatMap,
+                    _presenter.TypeMap,
                     runCallback: (run, classifiedText, position) =>
                     {
                         if (properties["Background"] is Brush highlightBrush)
@@ -136,7 +136,7 @@ namespace Microsoft.VisualStudio.LanguageServices.FindUsages
 
             private DisposableToolTip CreateDisposableToolTip(Document document, TextSpan sourceSpan)
             {
-                Presenter.AssertIsForeground();
+                _presenter.AssertIsForeground();
 
                 var controlService = document.Project.Solution.Workspace.Services.GetService<IContentControlService>();
                 var sourceText = document.GetTextSynchronously(CancellationToken.None);
@@ -144,7 +144,7 @@ namespace Microsoft.VisualStudio.LanguageServices.FindUsages
                 var excerptService = document.Services.GetService<IDocumentExcerptService>();
                 if (excerptService != null)
                 {
-                    var excerpt = Presenter.ThreadingContext.JoinableTaskFactory.Run(() => excerptService.TryExcerptAsync(document, sourceSpan, ExcerptMode.Tooltip, CancellationToken.None));
+                    var excerpt = _presenter.ThreadingContext.JoinableTaskFactory.Run(() => excerptService.TryExcerptAsync(document, sourceSpan, ExcerptMode.Tooltip, CancellationToken.None));
                     if (excerpt != null)
                     {
                         // get tooltip from excerpt service
